+++ core/common/src/main/java/alluxio/underfs/ObjectUnderFileSystem.java
@@ -233,6 +233,57 @@ public abstract class ObjectUnderFileSystem extends BaseUnderFileSystem {
     }
   }
 
+  @Override
+  public boolean renameDirectory(String src, String dst) throws IOException {
+    UnderFileStatus[] children = listInternal(src, false);
+    if (children == null) {
+      LOG.error("Failed to list directory {}, aborting rename.", src);
+      return false;
+    }
+    if (isFile(dst) || isDirectory(dst)) {
+      LOG.error("Unable to rename {} to {} because destination already exists.", src, dst);
+      return false;
+    }
+    // Source exists and is a file, and destination does not exist
+    // Rename the source folder first
+    if (!copy(convertToFolderName(src), convertToFolderName(dst))) {
+      return false;
+    }
+    // Rename each child in the src folder to destination/child
+    for (UnderFileStatus child : children) {
+      String childSrcPath = PathUtils.concatPath(src, child);
+      String childDstPath = PathUtils.concatPath(dst, child);
+      boolean success;
+      if (child.isDirectory()) {
+        // Recursive call
+        success = renameDirectory(childSrcPath, childDstPath);
+      } else {
+        success = renameFile(childSrcPath, childDstPath);
+      }
+      if (!success) {
+        LOG.error("Failed to rename path {}, aborting rename.", child);
+        return false;
+      }
+    }
+    // Delete src and everything under src
+    return deleteDirectory(src, DeleteOptions.defaults().setRecursive(true));
+  }
+
+  @Override
+  public boolean renameFile(String src, String dst) throws IOException {
+    if (!isFile(src)) {
+      LOG.error("Unable to rename {} to {} because source does not exist or is a directory.",
+          src, dst);
+      return false;
+    }
+    if (isFile(dst) || isDirectory(dst)) {
+      LOG.error("Unable to rename {} to {} because destination already exists.", src, dst);
+      return false;
+    }
+    // Source is a file and Destination does not exist
+    return copy(src, dst) && deleteInternal(src);
+  }
+
   // Default object UFS does not provide a mechanism for updating the configuration, no-op
   @Override
   public void setConf(Object conf) {}
@@ -263,6 +314,15 @@ public abstract class ObjectUnderFileSystem extends BaseUnderFileSystem {
     key = CommonUtils.stripSuffixIfPresent(key, PATH_SEPARATOR);
     return key + getFolderSuffix();
   }
+
+  /**
+   * Copies an object to another key.
+   *
+   * @param src the source key to copy
+   * @param dst the destination key to copy to
+   * @return true if the operation was successful, false otherwise
+   */
+  protected abstract boolean copy(String src, String dst) throws IOException;
 
   /**
    * Internal function to delete a key.
